add __hash__ for all classes comparing pointer. fixed the hash type to long.

+++ python/bindings/include/openravepy/openravepy_int.h
@@ -683,8 +683,8 @@ public:
     virtual py::object __unicode__() {
         return ConvertStringToUnicode(__str__());
     }
-    virtual int __hash__() {
-        return static_cast<int>(uintptr_t(_pbase.get()));
+    virtual long __hash__() {
+        return static_cast<long>(uintptr_t(_pbase.get()));
     }
     virtual bool __eq__(PyInterfaceBasePtr p) {
         return !!p && _pbase == p->GetInterfaceBase();
